backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests